:racehorse: [day15] Reduce memory footprint by using 32-bit integers

+++ include/aoc2020/day15.hpp
@@ -1,18 +1,19 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
 namespace aoc2020::day15 {
 
 struct memory_game {
-  size_t current_turn{1u};
-  size_t previous_turn_of_last_spoken_number{0u};
-  std::vector<size_t> number_to_previous_turn{};
+  uint32_t current_turn{1u};
+  uint32_t previous_turn_of_last_spoken_number{0u};
+  std::vector<uint32_t> number_to_previous_turn{};
 
-  memory_game(size_t turns) { number_to_previous_turn.resize(turns); }
+  memory_game(uint32_t turns) { number_to_previous_turn.resize(turns); }
 
-  auto speak(size_t number) {
+  auto speak(uint32_t number) {
     auto &previous_turn = number_to_previous_turn[number];
     previous_turn_of_last_spoken_number = previous_turn;
     previous_turn = current_turn++;
@@ -28,7 +29,7 @@ struct memory_game {
   }
 
   auto complete_all_turns() {
-    size_t last_spoken_number{};
+    uint32_t last_spoken_number{};
     while (current_turn <= std::size(number_to_previous_turn)) {
       last_spoken_number = speak_next();
     }
@@ -36,9 +37,9 @@ struct memory_game {
   }
 };
 
-auto play_game(auto &in, size_t turns) {
+auto play_game(auto &in, uint32_t turns) {
   memory_game game(turns);
-  size_t number;
+  uint32_t number;
   while (in >> number) {
     game.speak(number);
     in.get();
